Refactored some code
I have refactored some code to better follow standard and to be more clean, as well as updated some error messages and documentation.

+++ src/application.cpp
@@ -71,7 +71,7 @@ namespace Engine {
             camera.setPerspectiveProjection(glm::radians(50.0f), aspectRatio, 0.1f, 100.0f);
 
             if (auto commandBuffer = renderer.beginFrame()) {
-                size_t frameIndex = renderer.getCurrentFrameIndex();
+                int frameIndex = renderer.getCurrentFrameIndex();
                 FrameInfo frameInfo{frameIndex,
                                     deltaTime,
                                     commandBuffer,
