+++ src/utils/window/window.cpp
@@ -1,8 +1,7 @@
 #include "window.hpp"
 
 namespace Engine {
-    Window::Window(unsigned int width, unsigned int height, const char* title)
-    : width(width), height(height), window_title(title) {
+    Window::Window(int width, int height, const char* title) : width(width), height(height), window_title(title) {
         init();
     }
 
@@ -16,11 +15,7 @@ namespace Engine {
         glfwWindowHint(GLFW_CLIENT_API, GLFW_NO_API);
         glfwWindowHint(GLFW_RESIZABLE, GLFW_TRUE);
 
-        window = glfwCreateWindow(static_cast<int>(width),
-                                  static_cast<int>(height),
-                                  window_title,
-                                  nullptr,
-                                  nullptr);
+        window = glfwCreateWindow(width, height, window_title, nullptr, nullptr);
         glfwSetWindowUserPointer(window, this);
         glfwSetFramebufferSizeCallback(window, framebufferResizeCallback);
     }
@@ -35,7 +30,7 @@ namespace Engine {
     void Window::framebufferResizeCallback(GLFWwindow* window, int width, int height){
         auto win = reinterpret_cast<Window *>(glfwGetWindowUserPointer(window));
         win->framebufferResized = true;
-        win->width = static_cast<unsigned int>(width);
-        win->height = static_cast<unsigned int>(height);
+        win->width = width;
+        win->height = height;
     }
 }
